Add: `DialogBase` for Dialogs;
