+++ DnDCharCtor.ViewModels/IValidateableViewModel.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DnDCharCtor.ViewModels;
+
+public interface IValidateableViewModel
+{
+    bool HasValidationErrors { get; set; }
+
+    bool Validate();
+}
