+++ DnDCharCtor/DnDCharCtor.Ui/Components/Cards/CardBase.razor
@@ -6,10 +6,11 @@
     </FluentHeader>
 
     <FluentBodyContent>
-        <FluentStack Orientation="Orientation.Vertical"
-                     VerticalAlignment="VerticalAlignment.Center"
-                     HorizontalAlignment="HorizontalAlignment.Start"
-                     onclick="@EditCallback">
+        <FluentStack 
+            Orientation="Orientation.Vertical"
+            VerticalAlignment="VerticalAlignment.Center"
+            HorizontalAlignment="HorizontalAlignment.Start"
+            onclick="@EditCallback">
            @ChildContent
         </FluentStack>
