+++ DnDCharCtor/DnDCharCtor.Ui/DnDCharCtor.Ui.csproj
@@ -7,14 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="Components\InitialValidator.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Include="Components\InitialValidator.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
     <SupportedPlatform Include="browser" />
   </ItemGroup>
